GTP: Add config param for forcing a rule set (#109)
Our bots are only trained with Tromp-Taylor rules, but KGS doesn't have Tromp-Taylor rules. If a bot plays with non-Tromp-Taylor rules, then because the rules are features of the NN input, all NN inputs will be out-of-distribution. We therefore want to force our bots to always think they're playing with Tromp-Taylor rules.

This PR adds a config param ignoreGTPRuleChanges (in the same vein as the existing KataGo-raw config param ignoreGTPAndForceKomi) that makes KataGo ignore rule changes from katago-set-rule, katago-set-rules, kgs-rules, and loadsgf.