pytorchbackend: Note how to run GTP on small board sizes (#116)
If I `loadsgf` a 11x11 game in GTP mode with the Pytorch backend I get the error "Board len not yet supported", but it turns out there's an existing config parameter `gtpForceMaxNNSize` that can avoid this error. This PR notes in the code comment above the error that the user should set this config param.

Also modify .circleci to make GPU tests work again.